general clean up CompressExtension and subclasses
