+++ jetty-websocket/websocket-core/src/main/java/org/eclipse/jetty/websocket/core/internal/compress/CompressExtension.java
@@ -28,73 +28,43 @@ import org.eclipse.jetty.util.Callback;
 import org.eclipse.jetty.util.log.Log;
 import org.eclipse.jetty.util.log.Logger;
 import org.eclipse.jetty.websocket.core.AbstractExtension;
+import org.eclipse.jetty.websocket.core.BadPayloadException;
 import org.eclipse.jetty.websocket.core.Frame;
 import org.eclipse.jetty.websocket.core.MessageTooLargeException;
 import org.eclipse.jetty.websocket.core.OpCode;
+import org.eclipse.jetty.websocket.core.ProtocolException;
 import org.eclipse.jetty.websocket.core.internal.TransformingFlusher;
 
 public abstract class CompressExtension extends AbstractExtension
 {
-    protected static final byte[] TAIL_BYTES = new byte[]{0x00, 0x00, (byte)0xFF, (byte)0xFF};
-    protected static final ByteBuffer TAIL_BYTES_BUF = ByteBuffer.wrap(TAIL_BYTES);
+    private static final byte[] TAIL_BYTES = new byte[]{0x00, 0x00, (byte)0xFF, (byte)0xFF};
+    private static final ByteBuffer TAIL_BYTES_BUF = ByteBuffer.wrap(TAIL_BYTES);
     private static final Logger LOG = Log.getLogger(CompressExtension.class);
 
-    /**
-     * Never drop tail bytes 0000FFFF, from any frame type
-     */
-    protected static final int TAIL_DROP_NEVER = 0;
-    /**
-     * Always drop tail bytes 0000FFFF, from all frame types
-     */
-    protected static final int TAIL_DROP_ALWAYS = 1;
-    /**
-     * Only drop tail bytes 0000FFFF, from fin==true frames
-     */
-    protected static final int TAIL_DROP_FIN_ONLY = 2;
+    protected enum CompressionMode
+    {
+        FRAME,
+        MESSAGE
+    }
 
-    /**
-     * Always set RSV flag, on all frame types
-     */
-    protected static final int RSV_USE_ALWAYS = 0;
-    /**
-     * Only set RSV flag on first frame in multi-frame messages.
-     * <p>
-     * Note: this automatically means no-continuation frames have the RSV bit set
-     */
-    protected static final int RSV_USE_ONLY_FIRST = 1;
-
-    /**
-     * Inflater / Decompressed Buffer Size
-     */
-    protected static final int INFLATE_BUFFER_SIZE = 8 * 1024;
-
-    /**
-     * Deflater / Inflater: Maximum Input Buffer Size
-     */
-    protected static final int INPUT_MAX_BUFFER_SIZE = 8 * 1024;
-
-    /**
-     * Inflater : Output Buffer Size
-     */
+    private static final int COMPRESS_BUFFER_SIZE = 8 * 1024;
     private static final int DECOMPRESS_BUF_SIZE = 8 * 1024;
 
+    private final CompressionMode compressionMode;
     private final TransformingFlusher outgoingFlusher;
     private Deflater deflaterImpl;
     private Inflater inflaterImpl;
-    private int tailDrop;
-    private int rsvUse;
+    private boolean incomingCompressed;
 
     protected CompressExtension()
     {
-        tailDrop = getTailDropMode();
-        rsvUse = getRsvUseMode();
+        compressionMode = getCompressionMode();
 
         outgoingFlusher = new TransformingFlusher(getWebSocketCoreSession())
         {
             @Override
             protected boolean transform(Frame frame, Callback callback, boolean batch, boolean first)
             {
-                // Do not deflate control frames
                 if (OpCode.isControlFrame(frame.getOpCode()))
                 {
                     nextOutgoingFrame(frame, callback, batch);
@@ -106,54 +76,83 @@ public abstract class CompressExtension extends AbstractExtension
         };
     }
 
-    public Deflater getDeflater()
-    {
-        if (deflaterImpl == null)
-            deflaterImpl = getDeflaterPool().acquire();
-        return deflaterImpl;
-    }
-
-    public Inflater getInflater()
-    {
-        if (inflaterImpl == null)
-            inflaterImpl = getInflaterPool().acquire();
-        return inflaterImpl;
-    }
-
-    public void releaseInflater()
-    {
-        getInflaterPool().release(inflaterImpl);
-        inflaterImpl = null;
-    }
-
-    public void releaseDeflater()
-    {
-        getDeflaterPool().release(deflaterImpl);
-        deflaterImpl = null;
-    }
-
-    /**
-     * Indicates use of RSV1 flag for indicating deflation is in use.
-     */
     @Override
     public boolean isRsv1User()
     {
         return true;
     }
 
-    /**
-     * Return the mode of operation for dropping (or keeping) tail bytes in frames generated by compress (outgoing)
-     *
-     * @return either {@link #TAIL_DROP_ALWAYS}, {@link #TAIL_DROP_FIN_ONLY}, or {@link #TAIL_DROP_NEVER}
-     */
-    abstract int getTailDropMode();
+    abstract CompressionMode getCompressionMode();
 
-    /**
-     * Return the mode of operation for RSV flag use in frames generate by compress (outgoing)
-     *
-     * @return either {@link #RSV_USE_ALWAYS} or {@link #RSV_USE_ONLY_FIRST}
-     */
-    abstract int getRsvUseMode();
+    @Override
+    public void sendFrame(Frame frame, Callback callback, boolean batch)
+    {
+        // Compressed frames may increase in size so we need the flusher to fragment them.
+        outgoingFlusher.sendFrame(frame, callback, batch);
+    }
+
+    @Override
+    public void onFrame(Frame frame, Callback callback)
+    {
+        if (OpCode.isControlFrame(frame.getOpCode()))
+        {
+            nextIncomingFrame(frame, callback);
+            return;
+        }
+
+        switch (compressionMode)
+        {
+            case MESSAGE:
+                // This mode requires the RSV1 bit set only in the first frame.
+                // Subsequent continuation frames don't have RSV1 set, but are compressed.
+                switch (frame.getOpCode())
+                {
+                    case OpCode.TEXT:
+                    case OpCode.BINARY:
+                        incomingCompressed = frame.isRsv1();
+                        break;
+
+                    case OpCode.CONTINUATION:
+                        if (frame.isRsv1())
+                            callback.failed(new ProtocolException("Invalid RSV1 set on permessage-deflate CONTINUATION frame"));
+                        break;
+                    default:
+                        break;
+                }
+                break;
+
+            case FRAME:
+                incomingCompressed = frame.isRsv1();
+                break;
+
+            default:
+                throw new IllegalStateException();
+        }
+
+        if (!incomingCompressed)
+        {
+            nextIncomingFrame(frame, callback);
+            return;
+        }
+
+        ByteAccumulator accumulator = new ByteAccumulator(getWebSocketCoreSession().getMaxFrameSize());
+        try
+        {
+            ByteBuffer payload = frame.getPayload();
+            decompress(accumulator, payload);
+            if (frame.isFin() || compressionMode == CompressionMode.FRAME)
+                decompress(accumulator, TAIL_BYTES_BUF.slice());
+
+            forwardIncoming(frame, callback, accumulator);
+        }
+        catch (DataFormatException e)
+        {
+            throw new BadPayloadException(e);
+        }
+
+        if (frame.isFin())
+            incomingCompressed = false;
+    }
 
     protected void forwardIncoming(Frame frame, Callback callback, ByteAccumulator accumulator)
     {
@@ -218,9 +217,9 @@ public abstract class CompressExtension extends AbstractExtension
         ByteAccumulator accumulator = new ByteAccumulator(getWebSocketCoreSession().getMaxFrameSize());
         while (true)
         {
-            int bufferSize = DECOMPRESS_BUF_SIZE;
+            int bufferSize = COMPRESS_BUFFER_SIZE;
             if (maxFrameSize > 0)
-                bufferSize = (int)Math.min(maxFrameSize - accumulator.size(), DECOMPRESS_BUF_SIZE);
+                bufferSize = (int)Math.min(maxFrameSize - accumulator.size(), COMPRESS_BUFFER_SIZE);
 
             byte[] output = new byte[bufferSize];
             int compressed = deflater.deflate(output, 0, output.length, Deflater.SYNC_FLUSH);
@@ -247,26 +246,28 @@ public abstract class CompressExtension extends AbstractExtension
         if (payload.remaining() > 0)
         {
             // Handle tail bytes generated by SYNC_FLUSH.
-            if (LOG.isDebugEnabled())
-                LOG.debug("compressed[] bytes = {}", BufferUtil.toDetailString(payload));
+            switch (compressionMode)
+            {
+                case MESSAGE:
+                    if (finished && frame.isFin() && endsWithTail(payload))
+                    {
+                        payload.limit(payload.limit() - TAIL_BYTES.length);
+                        if (LOG.isDebugEnabled())
+                            LOG.debug("payload (TAIL_DROP_FIN_ONLY) = {}", BufferUtil.toDetailString(payload));
+                    }
+                    break;
 
-            if (tailDrop == TAIL_DROP_ALWAYS)
-            {
-                if (endsWithTail(payload))
-                {
-                    payload.limit(payload.limit() - TAIL_BYTES.length);
-                    if (LOG.isDebugEnabled())
-                        LOG.debug("payload (TAIL_DROP_ALWAYS) = {}", BufferUtil.toDetailString(payload));
-                }
-            }
-            else if (tailDrop == TAIL_DROP_FIN_ONLY)
-            {
-                if (finished && frame.isFin() && endsWithTail(payload))
-                {
-                    payload.limit(payload.limit() - TAIL_BYTES.length);
-                    if (LOG.isDebugEnabled())
-                        LOG.debug("payload (TAIL_DROP_FIN_ONLY) = {}", BufferUtil.toDetailString(payload));
-                }
+                case FRAME:
+                    if (endsWithTail(payload))
+                    {
+                        payload.limit(payload.limit() - TAIL_BYTES.length);
+                        if (LOG.isDebugEnabled())
+                            LOG.debug("payload (TAIL_DROP_ALWAYS) = {}", BufferUtil.toDetailString(payload));
+                    }
+                    break;
+
+                default:
+                    throw new IllegalStateException();
             }
         }
         else if (frame.isFin())
@@ -280,21 +281,12 @@ public abstract class CompressExtension extends AbstractExtension
             LOG.debug("Compressed {}: input:{} -> payload:{}", frame, outputLength, payload.remaining());
 
         Frame chunk = new Frame(first ? frame.getOpCode() : OpCode.CONTINUATION);
-        chunk.setRsv1(first || rsvUse != RSV_USE_ONLY_FIRST);
+        chunk.setRsv1(first || compressionMode != CompressionMode.FRAME);
         chunk.setPayload(payload);
         chunk.setFin(frame.isFin() && finished);
 
         nextOutgoingFrame(chunk, callback, batch);
         return finished;
-    }
-
-    @Override
-    public void sendFrame(Frame frame, Callback callback, boolean batch)
-    {
-        // We use a queue and an IteratingCallback to handle concurrency.
-        // We must compress and write atomically, otherwise the compression
-        // context on the other end gets confused.
-        outgoingFlusher.sendFrame(frame, callback, batch);
     }
 
     private static String toDetail(Inflater inflater)
@@ -324,6 +316,32 @@ public abstract class CompressExtension extends AbstractExtension
             }
         }
         return true;
+    }
+
+    public Deflater getDeflater()
+    {
+        if (deflaterImpl == null)
+            deflaterImpl = getDeflaterPool().acquire();
+        return deflaterImpl;
+    }
+
+    public Inflater getInflater()
+    {
+        if (inflaterImpl == null)
+            inflaterImpl = getInflaterPool().acquire();
+        return inflaterImpl;
+    }
+
+    public void releaseInflater()
+    {
+        getInflaterPool().release(inflaterImpl);
+        inflaterImpl = null;
+    }
+
+    public void releaseDeflater()
+    {
+        getDeflaterPool().release(deflaterImpl);
+        deflaterImpl = null;
     }
 
     @Override
