+++ jetty-websocket/websocket-core/src/main/java/org/eclipse/jetty/websocket/core/internal/compress/PerMessageDeflateExtension.java
@@ -18,19 +18,14 @@
 
 package org.eclipse.jetty.websocket.core.internal.compress;
 
-import java.nio.ByteBuffer;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.zip.DataFormatException;
 
 import org.eclipse.jetty.util.Callback;
 import org.eclipse.jetty.util.log.Log;
 import org.eclipse.jetty.util.log.Logger;
-import org.eclipse.jetty.websocket.core.BadPayloadException;
 import org.eclipse.jetty.websocket.core.ExtensionConfig;
 import org.eclipse.jetty.websocket.core.Frame;
-import org.eclipse.jetty.websocket.core.OpCode;
-import org.eclipse.jetty.websocket.core.ProtocolException;
 import org.eclipse.jetty.websocket.core.WebSocketComponents;
 
 /**
@@ -46,7 +41,6 @@ public class PerMessageDeflateExtension extends CompressExtension
     private ExtensionConfig configNegotiated;
     private boolean incomingContextTakeover = true;
     private boolean outgoingContextTakeover = true;
-    private boolean incomingCompressed;
 
     @Override
     public String getName()
@@ -55,52 +49,9 @@ public class PerMessageDeflateExtension extends CompressExtension
     }
 
     @Override
-    public void onFrame(Frame frame, Callback callback)
+    CompressionMode getCompressionMode()
     {
-        // Incoming frames are always non concurrent because
-        // they are read and parsed with a single thread, and
-        // therefore there is no need for synchronization.
-
-        // This extension requires the RSV1 bit set only in the first frame.
-        // Subsequent continuation frames don't have RSV1 set, but are compressed.
-        switch (frame.getOpCode())
-        {
-            case OpCode.TEXT:
-            case OpCode.BINARY:
-                incomingCompressed = frame.isRsv1();
-                break;
-
-            case OpCode.CONTINUATION:
-                if (frame.isRsv1())
-                    callback.failed(new ProtocolException("Invalid RSV1 set on permessage-deflate CONTINUATION frame"));
-                break;
-            default:
-                break;
-        }
-
-        if (OpCode.isControlFrame(frame.getOpCode()) || !incomingCompressed)
-        {
-            nextIncomingFrame(frame, callback);
-            return;
-        }
-
-        ByteAccumulator accumulator = new ByteAccumulator(getWebSocketCoreSession().getMaxFrameSize());
-        try
-        {
-            ByteBuffer payload = frame.getPayload();
-            decompress(accumulator, payload);
-            if (frame.isFin())
-                decompress(accumulator, TAIL_BYTES_BUF.slice());
-
-            forwardIncoming(frame, callback, accumulator);
-        }
-        catch (DataFormatException e)
-        {
-            throw new BadPayloadException(e);
-        }
-
-        if (frame.isFin())
-            incomingCompressed = false;
+        return CompressionMode.MESSAGE;
     }
 
     @Override
@@ -123,18 +74,6 @@ public class PerMessageDeflateExtension extends CompressExtension
             releaseDeflater();
         }
         super.nextOutgoingFrame(frame, callback, batch);
-    }
-
-    @Override
-    int getRsvUseMode()
-    {
-        return RSV_USE_ONLY_FIRST;
-    }
-
-    @Override
-    int getTailDropMode()
-    {
-        return TAIL_DROP_FIN_ONLY;
     }
 
     @Override
